pyproject.toml
[tool.poetry]
name = "sdet-ss-2024-practicum"
version = "0.1.0"
description = ""
authors = ["Capwell <proninc@yandex.ru>"]
readme = "README.md"
packages = [{include = "sdet_ss_2024_practicum"}]

[tool.poetry.dependencies]
python = "^3.9"
pytest = "^8.3.3"
selenium = "^4.24.0"
pytest-xdist = "^3.6.1"
allure-pytest = "^2.13.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.pytest.ini_options]
markers = [
    "ui: UI tests",    
]